Move Grid classes to single package
