+++ common/src/main/java/krabban91/kodvent/kodvent/utilities/grid/HyperCubeGrid.java
@@ -1,4 +1,7 @@
-package krabban91.kodvent.kodvent.utilities;
+package krabban91.kodvent.kodvent.utilities.grid;
+
+import krabban91.kodvent.kodvent.utilities.Point3D;
+import krabban91.kodvent.kodvent.utilities.TimePoint;
 
 import java.awt.*;
 import java.util.List;
